Move ability.ts to subfolder
