+++ biome.jsonc
@@ -32,6 +32,7 @@
       // TODO: these files are too big and complex, ignore them until their respective refactors
       "src/data/moves/move.ts",
       "src/data/ability.ts",
+      "src/data/abilities/ability.ts",
       "src/field/pokemon.ts",
 
       // this file is just too big:
