+++ src/data/abilities/ability.ts
@@ -1,29 +1,29 @@
-import type { EnemyPokemon, PokemonMove } from "../field/pokemon";
-import type Pokemon from "../field/pokemon";
-import { HitResult, MoveResult, PlayerPokemon } from "../field/pokemon";
+import type { EnemyPokemon, PokemonMove } from "../../field/pokemon";
+import type Pokemon from "../../field/pokemon";
+import { HitResult, MoveResult, PlayerPokemon } from "../../field/pokemon";
 import { PokemonType } from "#enums/pokemon-type";
 import { BooleanHolder, NumberHolder, toDmgValue, isNullOrUndefined, randSeedItem, randSeedInt, type Constructor } from "#app/utils";
-import { getPokemonNameWithAffix } from "../messages";
+import { getPokemonNameWithAffix } from "../../messages";
 import type { Weather } from "#app/data/weather";
-import type { BattlerTag } from "./battler-tags";
-import { BattlerTagLapseType, GroundedTag } from "./battler-tags";
+import type { BattlerTag } from "../battler-tags";
+import { BattlerTagLapseType, GroundedTag } from "../battler-tags";
 import { getNonVolatileStatusEffects, getStatusEffectDescriptor, getStatusEffectHealText } from "#app/data/status-effect";
-import { Gender } from "./gender";
-import type Move from "./moves/move";
-import { AttackMove, FlinchAttr, OneHitKOAttr, HitHealAttr, allMoves, StatusMove, SelfStatusMove, VariablePowerAttr, applyMoveAttrs, VariableMoveTypeAttr, RandomMovesetMoveAttr, RandomMoveAttr, NaturePowerAttr, CopyMoveAttr, NeutralDamageAgainstFlyingTypeMultiplierAttr, FixedDamageAttr } from "./moves/move";
+import { Gender } from "../gender";
+import type Move from "../moves/move";
+import { AttackMove, FlinchAttr, OneHitKOAttr, HitHealAttr, allMoves, StatusMove, SelfStatusMove, VariablePowerAttr, applyMoveAttrs, VariableMoveTypeAttr, RandomMovesetMoveAttr, RandomMoveAttr, NaturePowerAttr, CopyMoveAttr, NeutralDamageAgainstFlyingTypeMultiplierAttr, FixedDamageAttr } from "../moves/move";
 import { MoveFlags } from "#enums/MoveFlags";
 import { MoveTarget } from "#enums/MoveTarget";
 import { MoveCategory } from "#enums/MoveCategory";
-import type { ArenaTrapTag, SuppressAbilitiesTag } from "./arena-tag";
-import { ArenaTagSide } from "./arena-tag";
-import { BerryModifier, HitHealModifier, PokemonHeldItemModifier } from "../modifier/modifier";
-import { TerrainType } from "./terrain";
-import { SpeciesFormChangeAbilityTrigger, SpeciesFormChangeRevertWeatherFormTrigger, SpeciesFormChangeWeatherTrigger } from "./pokemon-forms";
+import type { ArenaTrapTag, SuppressAbilitiesTag } from "../arena-tag";
+import { ArenaTagSide } from "../arena-tag";
+import { BerryModifier, HitHealModifier, PokemonHeldItemModifier } from "../../modifier/modifier";
+import { TerrainType } from "../terrain";
+import { SpeciesFormChangeAbilityTrigger, SpeciesFormChangeRevertWeatherFormTrigger, SpeciesFormChangeWeatherTrigger } from "../pokemon-forms";
 import i18next from "i18next";
 import type { Localizable } from "#app/interfaces/locales";
-import { Command } from "../ui/command-ui-handler";
+import { Command } from "../../ui/command-ui-handler";
 import { BerryModifierType } from "#app/modifier/modifier-type";
-import { getPokeballName } from "./pokeball";
+import { getPokeballName } from "../pokeball";
 import type { BattlerIndex } from "#app/battle";
 import { BattleType } from "#app/battle";
 import { Abilities } from "#enums/abilities";
