+++ src/data/moves/move.ts
@@ -65,7 +65,7 @@ import {
   UserFieldMoveTypePowerBoostAbAttr,
   VariableMovePowerAbAttr,
   WonderSkinAbAttr,
-} from "../ability";
+} from "../abilities/ability";
 import {
   AttackTypeBoosterModifier,
   BerryModifier,
