+++ src/data/mystery-encounters/encounters/clowning-around-encounter.ts
@@ -38,7 +38,7 @@ import i18next from "i18next";
 import type { OptionSelectConfig } from "#app/ui/abstact-option-select-ui-handler";
 import type { PlayerPokemon } from "#app/field/pokemon";
 import { PokemonMove } from "#app/field/pokemon";
-import { Ability } from "#app/data/ability";
+import { Ability } from "#app/data/abilities/ability";
 import { BerryModifier } from "#app/modifier/modifier";
 import { BerryType } from "#enums/berry-type";
 import { BattlerIndex } from "#app/battle";
