+++ src/ui/starter-select-ui-handler.ts
@@ -8,8 +8,8 @@ import i18next from "i18next";
 import type BBCodeText from "phaser3-rex-plugins/plugins/bbcodetext";
 import { starterColors } from "#app/battle-scene";
 import { globalScene } from "#app/global-scene";
-import type { Ability } from "#app/data/ability";
-import { allAbilities } from "#app/data/ability";
+import type { Ability } from "#app/data/abilities/ability";
+import { allAbilities } from "#app/data/abilities/ability";
 import { speciesEggMoves } from "#app/data/balance/egg-moves";
 import { GrowthRate, getGrowthRateColor } from "#app/data/exp";
 import { Gender, getGenderColor, getGenderSymbol } from "#app/data/gender";
