+++ src/ui/summary-ui-handler.ts
@@ -31,7 +31,7 @@ import { loggedInUser } from "#app/account";
 import type { Variant } from "#app/sprites/variant";
 import { getVariantTint } from "#app/sprites/variant";
 import { Button } from "#enums/buttons";
-import type { Ability } from "#app/data/ability";
+import type { Ability } from "#app/data/abilities/ability";
 import i18next from "i18next";
 import { modifierSortFunc } from "#app/modifier/modifier";
 import { PlayerGender } from "#enums/player-gender";
