+++ test/abilities/friend_guard.test.ts
@@ -5,7 +5,7 @@ import GameManager from "#test/testUtils/gameManager";
 import Phaser from "phaser";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 import { BattlerIndex } from "#app/battle";
-import { allAbilities } from "#app/data/ability";
+import { allAbilities } from "#app/data/abilities/ability";
 import { allMoves } from "#app/data/moves/move";
 import { MoveCategory } from "#enums/MoveCategory";
