+++ test/abilities/magic_bounce.test.ts
@@ -1,5 +1,5 @@
 import { BattlerIndex } from "#app/battle";
-import { allAbilities } from "#app/data/ability";
+import { allAbilities } from "#app/data/abilities/ability";
 import { ArenaTagSide } from "#app/data/arena-tag";
 import { allMoves } from "#app/data/moves/move";
 import { ArenaTagType } from "#app/enums/arena-tag-type";
