+++ test/abilities/sand_veil.test.ts
@@ -1,4 +1,4 @@
-import { StatMultiplierAbAttr, allAbilities } from "#app/data/ability";
+import { StatMultiplierAbAttr, allAbilities } from "#app/data/abilities/ability";
 import { CommandPhase } from "#app/phases/command-phase";
 import { MoveEffectPhase } from "#app/phases/move-effect-phase";
 import { MoveEndPhase } from "#app/phases/move-end-phase";
