+++ test/abilities/steely_spirit.test.ts
@@ -1,4 +1,4 @@
-import { allAbilities } from "#app/data/ability";
+import { allAbilities } from "#app/data/abilities/ability";
 import { allMoves } from "#app/data/moves/move";
 import { Abilities } from "#app/enums/abilities";
 import { Moves } from "#enums/moves";
